feat(stm32): Implement STM32G474 170 MHz clock tree and monotonic time
-Configure the HSI16 PLL and propagate peripheral kernel clocks
-Extend hal_micros64 across the 32-bit wrap in bare-metal and FreeRTOS
-Add timing tests and document hardware validation

+++ src/hal/impl/stm32g474/hal_rgb_led.cpp
@@ -66,9 +66,11 @@ bool jh_hal_rgb_led_write_pixels(const uint8_t *pixels, uint32_t num_bytes,
   cycles_enable();
   const uint32_t bit_cycles = JH_G474_CORE_CLOCK_HZ / 800000u;
   const uint32_t zero_high_cycles =
-      ((JH_G474_CORE_CLOCK_HZ * 35u) + 50000000u) / 100000000u;
+      (uint32_t)((((uint64_t)JH_G474_CORE_CLOCK_HZ * 35u) + 50000000u) /
+                 100000000u);
   const uint32_t one_high_cycles =
-      ((JH_G474_CORE_CLOCK_HZ * 80u) + 50000000u) / 100000000u;
+      (uint32_t)((((uint64_t)JH_G474_CORE_CLOCK_HZ * 80u) + 50000000u) /
+                 100000000u);
   if (zero_high_cycles == 0u ||
       !(zero_high_cycles < one_high_cycles && one_high_cycles < bit_cycles)) {
     return false;
